fix the hparameter dtype error

+++ intel_extension_for_transformers/llm/runtime/graph/models/model_utils/model_types.h
@@ -135,8 +135,8 @@ struct model_hparams {
 
   // ChatGLM-1
   int32_t inner_hidden_size = 0;
-  int32_t rope_scaling_factor = 0;
-  float original_max_position_embeddings = 0.0f;
+  float rope_scaling_factor = 0.0f;
+  int32_t original_max_position_embeddings = 0;
   int32_t use_yarn = 0;
 
   bool operator!=(const model_hparams& other) const {
